NavigationRailの構成の修正
