+++ layouts/default.vue
@@ -1,6 +1,34 @@
+<script setup lang="ts">
+import { Compass, Feather, History, Settings2, User, Users } from 'lucide-vue-next'
+
+const currentUser = useCurrentUserStore()
+const linkContents = ref([
+  { to: '/', icon: History, text: 'アプデ' },
+  { to: '/browse', icon: Compass, text: '見つける' },
+  { to: `/u/${currentUser.name}`, icon: User, text: '君とは' },
+  { to: '/friends', icon: Users, text: '知り合い' },
+  { to: '/settings', icon: Settings2, text: 'せってー' }
+])
+</script>
+
 <template>
   <div class="layout">
-    <OrgNavigationRail />
+    <MaterialNavigationRail>
+      <template #fab>
+        <MaterialFAB>
+          <Feather />
+        </MaterialFAB>
+      </template>
+      <template #link-list>
+        <MaterialNavigationRailLink
+          v-for="linkContent in linkContents"
+          :key="linkContent.to"
+          :to="linkContent.to"
+          :label="linkContent.text"
+          :icon="linkContent.icon"
+        />
+      </template>
+    </MaterialNavigationRail>
     <main>
       <slot />
     </main>
